+++ pkg/usecase/vendingservice_test.go
@@ -1,4 +1,4 @@
-package service
+package usecase
 
 import (
 	"context"
